replicaCodice/test_SAVIA/retriever/requirements.txt
accelerate==1.0.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
anaconda-anon-usage==0.4.4
annotated-types==0.7.0
anyio==4.6.2.post1
archspec==0.2.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.0.5
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
boltons==23.0.0
Brotli==1.0.9
certifi==2024.2.2
cffi==1.16.0
chardet==4.0.0
charset-normalizer==2.0.4
click==8.1.7
coloredlogs==15.0.1
comm==0.2.2
conda==23.5.2
conda-build==24.3.0
conda-content-trust==0.2.0
conda_index==0.4.0
conda-libmamba-solver==23.7.0
conda-package-handling==2.2.0
conda_package_streaming==0.9.0
cryptography==42.0.5
dataclasses-json==0.6.7
datasets==3.0.1
debugpy==1.8.7
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.8
dirtyjson==1.0.8
distro==1.8.0
dnspython==2.6.1
exceptiongroup==1.2.0
executing==0.8.3
expecttest==0.2.1
faiss-gpu==1.7.2
fastjsonschema==2.20.0
filelock==3.13.1
FlagEmbedding==1.2.11
Flask==3.0.3
flatbuffers==24.3.25
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.3.1
gliner==0.2.13
gmpy2==2.1.2
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.26.1
humanfriendly==10.0
hypothesis==6.100.1
idna==3.4
ipykernel==6.29.5
ipython==8.20.0
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.18.1
Jinja2==3.1.3
jiter==0.6.1
joblib==1.4.2
json5==0.9.25
jsonpatch==1.33
jsonpointer==2.1
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter==1.1.1
jupyter_client==8.6.3
jupyter-console==6.6.3
jupyter_core==5.7.2
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
lark==1.1.9
libarchive-c==2.9
libmambapy==1.5.8
llama-cloud==0.1.4
llama-index==0.11.19
llama-index-agent-openai==0.3.4
llama-index-cli==0.3.1
llama-index-core==0.11.19
llama-index-embeddings-openai==0.2.5
llama-index-indices-managed-llama-cloud==0.4.0
llama-index-legacy==0.9.48.post3
llama-index-llms-openai==0.2.15
llama-index-multi-modal-llms-openai==0.2.2
llama-index-program-openai==0.2.0
llama-index-question-gen-openai==0.2.0
llama-index-readers-file==0.2.2
llama-index-readers-llama-parse==0.3.0
llama-parse==0.5.10
MarkupSafe==2.1.3
marshmallow==3.23.0
matplotlib-inline==0.1.6
menuinst==2.0.2
mistune==3.0.2
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
more-itertools==10.1.0
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.1
nltk==3.9.1
notebook==7.2.2
notebook_shim==0.2.4
numpy==1.26.4
onnxruntime==1.19.2
openai==1.52.0
optree==0.11.0
overrides==7.7.0
packaging==23.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.3
peft==0.13.2
pexpect==4.8.0
pillow==10.2.0
pip==23.3.1
pkginfo==1.9.6
platformdirs==3.10.0
pluggy==1.0.0
prometheus_client==0.21.0
prompt-toolkit==3.0.43
propcache==0.2.0
protobuf==5.28.2
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==17.0.0
pycosat==0.6.6
pycparser==2.21
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.15.1
pymongo==4.10.1
pyOpenSSL==24.0.0
pypdf==4.3.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-etcd==0.4.5
python-json-logger==2.0.7
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==26.2.0
rank-bm25==0.2.2
referencing==0.30.2
regex==2024.9.11
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.10.6
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
Send2Trash==1.8.3
sentence-transformers==3.2.1
sentencepiece==0.2.0
setuptools==68.2.2
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.36
stack-data==0.2.0
striprtf==0.0.26
sympy==1.12
tenacity==8.5.0
terminado==0.18.1
threadpoolctl==3.5.0
tiktoken==0.8.0
tinycss2==1.3.0
tokenizers==0.20.1
tomli==2.0.1
toolz==0.12.0
torch==2.3.0
torchaudio==2.3.0
torchelastic==0.2.2
torchvision==0.18.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.7.1
transformers==4.45.2
triton==2.3.0
truststore==0.8.0
types-dataclasses==0.6.6
types-python-dateutil==2.9.0.20241003
typing_extensions==4.11.0
typing-inspect==0.9.0
tzdata==2024.2
uri-template==1.3.0
urllib3==2.1.0
wcwidth==0.2.5
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.4
wheel==0.41.2
widgetsnbextension==4.0.13
wrapt==1.16.0
xxhash==3.5.0
yarl==1.15.5
zstandard==0.19.0
